task: preprocessed valid coupons util
